+++ CHANGELOG.md
@@ -21,7 +21,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - Added index comparison methods to `PhmmIndex` (behind `dev-phmm` feature gate)
 - Condensed the `PhmmParamKind` variants in the `visit_params` API (behind `dev-phmm` feature gate)
 - Added row and column accessors to `TransitionParams` (behind `dev-phmm` and `alignment-diagnostics` feature gates)
-- Added `WithErrorContext` trait and `ErrorWithContext` struct for updated error-handling.
+- Added `WithErrorContext` and `ResultWithErrorContext` traits using the `ErrorWithContext` struct for updated error-handling.
 
 ### Changed